requirements.txt
numpy
scipy
matplotlib
pandas==1.2.0
jax==0.20
uproot==3.12.0
rootpy==1.0.1
uncertainties
jinja2
tabulate